+++ molecularnodes/entities/trajectory/__init__.py
@@ -1,6 +1,7 @@
 from . import selections
 from . import ui
+from . import props
 from .ui import load
 from .trajectory import Trajectory
 
-CLASSES = selections.CLASSSES + ui.CLASSES
+CLASSES = props.CLASSSES + ui.CLASSES
