+++ molecularnodes/ui/__init__.py
@@ -1,3 +1,4 @@
-from . import panel, node_menu
+from .panel import CLASSES as PANEL_CLASSES
+from .node_menu import CLASSES as NODE_MENU_CLASSES
 
-CLASSES = panel.CLASSES + node_menu.CLASSES
+CLASSES = PANEL_CLASSES + NODE_MENU_CLASSES
